+++ libs/my_c/src/list/my_list_to_array.c
@@ -20,13 +20,13 @@
  */
 void **my_list_to_array(node_t *list)
 {
-    int size = my_list_size(list);
+    size_t size = my_list_size(list);
     void **array = my_calloc(size + 1, sizeof(void *));
     node_t *tmp = list;
 
     if (array == NULL)
         return NULL;
-    for (int i = 0; tmp != NULL; i++) {
+    for (size_t i = 0; tmp != NULL; i++) {
         array[i] = tmp->data;
         tmp = tmp->next;
     }
