+++ libs/my_c/src/string/my_strlowcase.c
@@ -20,7 +20,7 @@
  */
 char *my_strlowcase(char *str)
 {
-    for (int i = 0; str[i] != '\0'; i++) {
+    for (size_t i = 0; str[i] != '\0'; i++) {
         if (str[i] >= 65 && str[i] <= 90)
             str[i] = str[i] + 32;
     }
